test(editor): add tests for multi workspace folder setup (#10904)
We have now 3 different test setups:

###  `test()`
Will be executed in both modes.  Should be used to make sure this works for both ways.

###  `testSingleFolderMode()`
Will only be executed for single folders. Testing with multi folder is redundant.

###  `testMultiFolderMode()`
Will only be executed in multi folder mode. These are special tests which are only designed for multi folders and makes no sense that it will work in single folder.

---

`loadFixture()` and `getDiagnostics()` will use the second folder when the multi folder mode is activated
