+++ editors/vscode/fixtures/debugger_error/.oxlintrc.json
@@ -1,0 +1,5 @@
+{
+  "rules": {
+    "no-debugger": "error"
+  }
+}
