+++ editors/vscode/fixtures/debugger_error/debugger.js
@@ -1,0 +1,1 @@
+/* 😊 */debugger;
